+++ __init__.py
@@ -93,6 +93,80 @@
 ### Added
 ## - Display Log Info (is missing in UI)
 
+## v0.1.2
+## 2019-09-11
+### Added
+## - 3D_view save location
+
+## v0.1.3
+## 2019-11-11
+### Added
+## - Sync View Settings
+
+## v0.1.4
+## 2019-12-20
+### Added
+## - Save preset
+
+## v0.1.5
+## 2020-01-15
+### Fixed
+## - Viewport render animation operator was missing part
+
+## v0.1.6
+## 2020-03-11
+### Fixed
+## - Render Display changed named
+##
+### Changed
+## - Moved Render in to render section
+
+## v0.1.7
+## 2020-03-17
+### Fixed
+## - Space in enumproperty
+
+## v0.1.8
+## 2020-03-20
+### Fixed
+## - Warning error Panel class
+
+## v0.1.9
+## 2021-02-26
+### Added
+## - Show console (Windows)
+## - Viewport Render Frames
+
+## v0.1.9
+## 2021-03-05
+### Added
+## - Operator to go to file location > missing since old blender
+## 
+
+## v0.2.0
+## 2021-08-05
+### Fixed
+## - Menu operator for console > hode for OSX
+## 
+
+## v0.2.1
+## 2022-01-20
+### Fixed
+## - Store view after looking at MACHIN3 method. I was close > Set store view as a default
+## - If workspace are not in default names, revert to a default name and interaction mode > needs work for keymaps
+
+## v0.2.1
+## 2022-01-21
+### Changed
+## - Try to save names workspaces, issue is still present when name is changed. Workspaces dont have any info order
+
+"""
+
+	TODO
+	- When adjusting workarea name, replace stored item with new item name
+	^ We need to store the screenspace(orden in order so we can easily check them), run a check if false replace with new name
+	- Expand to order workflows: greasepencil, see other starting templates
+"""
 #######################################################
 
 bl_info = {
@@ -100,41 +174,94 @@ bl_info = {
 	"description": "QuickSwitch is a little helper to make it easier to switch render engines & workspaces",
 	"location": "3D VIEW > Quick Switch (see hotkeys)",
 	"author": "Rombout Versluijs",
-	"version": (0, 1, 1),
+	"version": (0, 2, 2),
 	"blender": (2, 80, 0),
 	"wiki_url": "https://github.com/schroef/quickswitch",
 	"tracker_url": "https://github.com/schroef/quickswitch/issues",
 	"category": "Viewport"
 }
 
-import bpy
+import bpy, sys, os, subprocess
 import rna_keymap_ui
-#from bl_operators.presets import AddPresetBase, PresetMenu
+from sys import platform
 from bpy.app.handlers import persistent
-#from . import AddPresetBase
+from bpy_extras.io_utils import ImportHelper
+# from bl_operators.presets import AddPresetBase, PresetMenu
+# from . import AddPresetBase
 
 from bpy.types import (
 	Panel, WindowManager, AddonPreferences, Menu, Operator, Scene, PropertyGroup
 	)
 from bpy.props import (
-	EnumProperty, StringProperty, BoolProperty, IntProperty, PointerProperty
+	EnumProperty, StringProperty, BoolProperty, IntProperty, PointerProperty, FloatProperty, FloatVectorProperty
 	)
 
 
-def avail_workspaces(self,context):
+def get_names_workspaces(self,context):
 	'''
 	enumerates available Workspaces and adding more items:
-	Maximize Area - will toggle current area to maximum window size
-	User Preferences - opens User Preferences window
+	store it using id
+	https://blender.stackexchange.com/questions/78133/dynamic-enumproperty-values-changing-unexpectedly
 	'''
+	ws = bpy.data.workspaces
+	items_store = []
+	wsNames = []
+	wsNames.append(("Preferences","Preferences","Preferences", 0))
+	#Scan the list of IDs to see if we already have one for this mesh
+	maxid = -1
+	id = -1
+	found = False
+	for ws in bpy.data.workspaces:
+		for idrec in items_store:
+			# print("idrec %s - %s" % (idrec[0],idrec[1]))
+			id = idrec[0]
+			if id > maxid:
+				maxid = id
+			if idrec[1] == ws.name:
+				found = True
+				break
+		if not found:
+			items_store.append((maxid+1, ws.name,ws.name))
+		# AMENDED CODE - include the ID	
+		# print("ws %s - %s" % (ws.name, id))
+		wsNames.append( (ws.name, ws.name, ws.name, id) )
+	# print(wsNames)
+	return wsNames
 
-	screen = bpy.data.workspaces
-	screens = [ ("Preferences", "Preferences", "Preferences")]
-			# (identifier, name, description) optionally: (.., icon name, unique number)
+	
+	# wsNames = [ ("Preferences", "Preferences", "", 0)]
+	# 		# (identifier, name, description) optionally: (.., icon name, unique number)
 
-	for screen in bpy.data.workspaces:
-		screens.append((screen.name, screen.name, screen.name))
-	return screens
+	# i = 1
+	# for ws in bpy.data.workspaces:
+	# 	wsNames.append((ws.name, ws.name, "", i))
+	# 	i+=1
+	# return wsNames
+
+# def get_names_workspaces(self,context):
+# 	'''
+# 	enumerates available Workspaces and adding more items:
+# 	Maximize Area - will toggle current area to maximum window size
+# 	User Preferences - opens User Preferences window
+# 	'''
+
+# 	ws = bpy.data.workspaces
+# 	wsNames = [ ("Preferences", "Preferences", "", 0)]
+# 			# (identifier, name, description) optionally: (.., icon name, unique number)
+
+# 	i = 1
+# 	for ws in bpy.data.workspaces:
+# 		wsNames.append((ws.name, ws.name, "", i))
+# 		i+=1
+# 	return wsNames
+
+
+def update_names_workspaces(self,context):
+	'''
+	check order of workspace and see what has changed since last call
+	'''
+	print(self)
+
 
 
 #def defaultWSSmodes(self, context):
@@ -152,55 +279,70 @@ def avail_workspaces(self,context):
 #		i+=1
 #	return wsModes
 
+
+class QS_workspaceStore(PropertyGroup):
+	qsWS0 : StringProperty(
+		name = "ws0",
+		default='Layout')
+
+
 class QS_defaultWSSmodes(PropertyGroup):
-	Animation : StringProperty(
-		name = "Animation",
-		default='POSE')
-	Compositing : StringProperty(
-		name = "Compositing",
-		default='OBJECT')
 	Layout : StringProperty(
 		name = "Layout",
 		default='OBJECT')
 	Modeling : StringProperty(
 		name = "Modeling",
 		default='EDIT')
+	Sculpting : StringProperty(
+		name = "Sculpting",
+		default='SCULPT')
+	UV_Editing : StringProperty(
+		name = "UV_Editing",
+		default='EDIT')
+	Texture_Paint : StringProperty(
+		name = "Texture_Paint",
+		default='TEXTURE_PAINT')
+	Shading : StringProperty(
+		name = "Shading",
+		default='OBJECT')
 	Rendering : StringProperty(
 		name = "Rendering",
+		default='OBJECT')
+	Animation : StringProperty(
+		name = "Animation",
+		default='POSE')
+	Compositing : StringProperty(
+		name = "Compositing",
 		default='OBJECT')
 	Scripting : StringProperty(
 		name = "Scripting",
 		default='OBJECT')
-	Sculpting : StringProperty(
-		name = "Sculpting",
-		default='SCULPT')
-	Shading : StringProperty(
-		name = "Shading",
+	Default : StringProperty(
+		name = "Default",
 		default='OBJECT')
-	Texture_Paint : StringProperty(
-		name = "Texture_Paint",
-		default='TEXTURE_PAINT')
-	UV_Editing : StringProperty(
-		name = "UV_Editing",
-		default='EDIT')
 
 @persistent
 def on_scene_update(scene):
 	context = bpy.context
 	scene = context.scene
 	ws = context.workspace
+	# settings = scene["qsWSsmode"]
 	settings = scene.qsWSsmode
-
-	settings.Animation = 'POSE'
-	settings.Compositing = 'OBJECT'
+	wsstore = scene.qsWSstore
+	wsstore.qsWS0 = bpy.data.workspaces['Modeling'].object_mode
+	# print(wsstore.qsWS0)
 	settings.Layout = 'OBJECT'
-	settings.Modeling = bpy.data.workspaces['Modeling'].object_mode
-	settings.Rendering = 'OBJECT'
-	settings.Scripting = 'OBJECT'
+	# settings.Modeling = bpy.data.workspaces['Modeling'].object_mode
+	settings.Modeling = 'EDIT'
 	settings.Sculpting = 'SCULPT'
-	settings.Shading = 'OBJECT'
-	settings.Texture_Paint = 'TEXTURE_PAINT'
 	settings.UV_Editing = 'EDIT'
+	settings.Texture_Paint = 'TEXTURE_PAINT'
+	settings.Shading = 'OBJECT'
+	settings.Animation = 'POSE'
+	settings.Rendering = 'OBJECT'
+	settings.Compositing = 'OBJECT'
+	settings.Scripting = 'OBJECT'
+	settings.Default = 'OBJECT'
 
 #class SetWSmodes(Operator):
 #	bl_idname = "qs.set_ws_modes"
@@ -222,129 +364,131 @@ distance = 0
 matrix = (0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
 rotation = (0,0,0,0)
 #bpy.types.Scene.qsDefWSmodes = bpy.props.EnumProperty(name = "Def WorkSpace Modes", items=defaultWSSmodes)
-bpy.types.Scene.qsStore3dView = bpy.props.BoolProperty(name = "Store View", default=False, description="This stores the view in 3D View and sets this view to the workspace you are moving to.")
+bpy.types.Scene.qsStore3dView = bpy.props.BoolProperty(name = "Store View", default=True, description="This stores the view in 3D View and sets this view to the workspace you are moving to.")
 
 bpy.types.Scene.qsKeepMode = bpy.props.BoolProperty(name = "Keep Mode", default=False, description="This stores the current Object Interaction Mode, now when you switch workspaces you will return to prior mode and not the default one.")
 
-class QS_StoreView(object):
-	#"""Get current view in the 3D View and store it for reuse"""
-	#bl_idname = "qs.store_view"
-	#bl_label = "Store View"
+# class QS_StoreView():
+# 	## https://stackoverflow.com/questions/9028398/change-viewport-angle-in-blender-using-python
+# 	#"""Get current view in the 3D View and store it for reuse"""
+# 	#bl_idname = "qs.store_view"
+# 	#bl_label = "Store View"
 
-	def __init__(self, context):
-		self.context = context
-		#self.ViewPort = self.ViewPort()
+# 	def __init__(self, context):
+# 		self.context = context
+# 		#self.ViewPort = self.ViewPort()
 
-	@property
-	def view(self):
-		""" Returns the set of 3D views.
-		"""
-		rtn = []
-		for a in self.context.window.screen.areas:
-			if a.type == 'VIEW_3D':
-				rtn.append(a)
-		return rtn
+# 	@property
+# 	def view(self, context):
+# 		""" Returns the set of 3D views.
+# 		"""
+# 		rtn = []
+# 		# for a in self.context.window.screen.areas:
+# 		for a in context.area:
+# 			if a.type == 'VIEW_3D':
+# 				rtn.append(a)
+# 		return rtn
 
-	def ViewPort(self, context):
-		""" Return position, rotation data about a given view for the first space attached to it """
-		global viewLoc, distance, rotation
+# 	def ViewPort(self, context):
+# 		""" Return position, rotation data about a given view for the first space attached to it """
+# 		global viewLoc, distance, rotation
 
-		for area in bpy.context.screen.areas:
-			if area.type == 'VIEW_3D':
+# 		# for area in bpy.context.screen.areas:
+# 		for area in context.area:
+# 			if area.type == 'VIEW_3D':
 
-				rv3d = area.spaces[0].region_3d
-				view_Loc = rv3d.view_location
-				viewloc = rv3d.view_location
-				#viewloc = view_Loc[0],view_Loc[1],view_Loc[2]
-				distance = rv3d.view_distance
-				matrix = rv3d.view_matrix
-				print(distance)
-				#camera_pos = self.camera_position(matrix)
-				rotation = rv3d.view_rotation
-				#rotation = rotation[0],rotation[1],rotation[2],rotation[3]
-				#return view_Loc, rotation, distance #camera_pos,
-		return
+# 				rv3d = area.spaces[0].region_3d
+# 				view_Loc = rv3d.view_location
+# 				viewloc = rv3d.view_location
+# 				#viewloc = view_Loc[0],view_Loc[1],view_Loc[2]
+# 				distance = rv3d.view_distance
+# 				matrix = rv3d.view_matrix
+# 				print(distance, "distannce")
+# 				#camera_pos = self.camera_position(matrix)
+# 				rotation = rv3d.view_rotation
+# 				#rotation = rotation[0],rotation[1],rotation[2],rotation[3]
+# 				#return view_Loc, rotation, distance #camera_pos,
+# 		return
 
 def ViewPort(wsn):
 		""" Return position, rotation data about a given view for the first space attached to it """
 		global viewLoc, distance, matrix, rotation
 
 		for i, area in enumerate(bpy.context.screen.areas):
-		#	print(bpy.context.window.workspace.name)
-		#for area in bpy.context.window.workspace[bpy.data.workspaces[0]].screen.areas:
 			if area.type == 'VIEW_3D':
-				#print(bpy.context.window.workspace[bpy.data.workspaces[wsn]])
-				#print(bpy.data.workspaces[wsn].screen.areas)
 				wsp = bpy.context.window.workspace
-				print(wsp.screen.areas)
 				rv3d = area.spaces[0].region_3d
-				#view_Loc = rv3d.view_location
 				viewLoc = rv3d.view_location
-				#viewloc = view_Loc[0],view_Loc[1],view_Loc[2]
 				distance = rv3d.view_distance
 				matrix = rv3d.view_matrix
-				print(distance)
-				#camera_pos = self.camera_position(matrix)
 				rotation = rv3d.view_rotation
+				#camera_pos = self.camera_position(matrix)
 				#rotation = rotation[0],rotation[1],rotation[2],rotation[3]
 				#return view_Loc, rotation, distance #camera_pos,
+				# print("Vloc: %s | Dist %s | Mtrx %s | Rot %s" % (viewLoc, distance, matrix, rotation))
 
-#class QS_Store3DView(PropertyGroup):
-#	viewLoc : StringProperty(
-#		name = "viewLoc",
-#		default="0,0,0")
-#
-#	distance : StringProperty(
-#		name = "distance",
-#		default="0")
-#
-#	rotation : StringProperty(
-#		name = "rotation",
-#		default="0,0,0,0")
+
+class QS_Store3DView(PropertyGroup):
+	distance: FloatProperty()
+	viewLoc: FloatVectorProperty (
+		subtype='TRANSLATION')
+	rotation: FloatVectorProperty (
+		subtype='QUATERNION',
+		size=4)
+	name: StringProperty()
+
 
 @persistent
 def on_ws_switch(scene):
 	context = bpy.context
 	scene = context.scene
-	settings = scene.qsStoreView
+	qssv = scene.qsStoreView
 
-	#viewport = QS_StoreView.ViewPort(context, context)
-	#settings.viewLoc =str(viewLoc)
-	#settings.distance =str(distance)
-	#settings.rotation =str(rotation)
-	global viewLoc, distance, matrix, rotation
 	wsn = context.workspace
 	#try:
 	ViewPort(wsn)
-	viewLoc = viewLoc
-	distance = distance
-	matrix = matrix
-	rotation = rotation
 
-	print("## %s - %s -%s" % (viewLoc,distance, rotation))
+	global viewLoc, distance, matrix, rotation
+	# viewport = QS_StoreView.ViewPort(context, context)
+	qssv.viewLoc = viewLoc#str(viewLoc)
+	qssv.distance = distance#str(distance)
+	qssv.rotation = rotation#str(rotation)
+
+	# viewLoc = viewLoc
+	# distance = distance
+	# matrix = matrix 
+	# rotation = rotation
+
+	# print("## %s - %s -%s" % (viewLoc,distance, rotation))
 	#except:
 	#	pass
+	getView(context)
 
 
 addon_keymaps = []
 
 def add_hotkey():
-	preferences = bpy.context.preferences
-	addon_prefs = preferences.addons[__name__].preferences
+	# preferences = bpy.context.preferences
+	# addon_prefs = preferences.addons[__name__].preferences
 
+	# Use alt or cmd for windows and osx
+	cmdK = False if platform == "win32" else True
+	altK = True if platform == "win32" else False
+
+	# print("QS hotkeys added")
 	wm = bpy.context.window_manager
 
 	kc = wm.keyconfigs.addon    # for hotkeys within an addon
 	km = kc.keymaps.new(name = "Screen", space_type = "EMPTY")
 
 	#Add quick menu
-	kmi = km.keymap_items.new("wm.call_menu",  value='PRESS', type='W', ctrl=False, alt=True, shift=False, oskey=False)
+	kmi = km.keymap_items.new("wm.call_menu",  value='PRESS', type='Q', ctrl=False, alt=altK, shift=False, oskey=cmdK)
 	kmi.properties.name = "QS_MT_WorkspaceSwitchMenu"
 	kmi.active = True
 	addon_keymaps.append((km, kmi))
 
 	#Add quick pie menu
-	kmi = km.keymap_items.new("wm.call_menu_pie",  value='PRESS', type='W', ctrl=False, alt=True, shift=True, oskey=False)
+	kmi = km.keymap_items.new("wm.call_menu_pie",  value='PRESS', type='Q', ctrl=False, alt=altK, shift=True, oskey=cmdK)
 	kmi.properties.name = "QS_MT_WorkspaceSwitchPieMenu"
 	kmi.active = True
 	addon_keymaps.append((km, kmi))
@@ -353,15 +497,16 @@ def add_hotkey():
 	wm = bpy.context.window_manager
 	kc = wm.keyconfigs.addon
 	km = kc.keymaps.new(name = "Screen", space_type = "EMPTY")
-	kmi = km.keymap_items.new("wm.call_menu", value='PRESS', type='E', alt=True, shift=True)
+	kmi = km.keymap_items.new("wm.call_menu", value='PRESS', type='E', alt=altK, shift=True, oskey=cmdK)
 	kmi.properties.name = "QS_MT_QuickSwitchEngine"
 	kmi.active = True
 	addon_keymaps.append((km, kmi))
 
 	hKeys = [("NUMPAD_1"), ("NUMPAD_2"), ("NUMPAD_3"), ("NUMPAD_4"),("NUMPAD_5"),("NUMPAD_6"),("NUMPAD_7"),("NUMPAD_8"),("NUMPAD_9"),("NUMPAD_0")]
 	i=0
+	
 	for screen in hKeys:
-		kmi = km.keymap_items.new("qs.workspace_set_layout", hKeys[i], "PRESS", oskey=True)  # ...and if not found then add it
+		kmi = km.keymap_items.new("qs.workspace_set_layout", hKeys[i], "PRESS", alt=altK, oskey=cmdK)  # ...and if not found then add it
 		kmi.properties.layoutName = "WorkspaceSwitcher"+str(i)  # also set proper name
 		kmi.active = True
 		addon_keymaps.append((km, kmi)) # also append to global (addon level) hotkey list for easy management
@@ -374,7 +519,7 @@ def add_hotkey():
 #	preset_menu = "QS_PT_presets"
 #
 #	preset_defines = [
-#		"qs = bpy.context.cloth"
+#		"qs = bpy.context.scene.quickswit"
 #	]
 #
 #	preset_values = [
@@ -436,6 +581,8 @@ class QS_MT_QuickSwitchEngine(Menu):
 	def draw(self, context):
 		layout = self.layout
 		layout.operator_context = "INVOKE_AREA"
+		prefs = context.preferences
+		view = prefs.view
 
 		scene = context.scene
 		rd = scene.render
@@ -452,22 +599,30 @@ class QS_MT_QuickSwitchEngine(Menu):
 		props = layout.operator("render.render", text="Render Animation", icon='RENDER_ANIMATION')
 		props.animation = True
 		props.use_viewport = True
+		layout.prop_menu_enum(view, "render_display_type", text="Render in")
 
 		layout.separator()
-
 		layout.operator("render.view_show", text="View Render")
-
 		layout.operator("render.play_rendered_anim", text="View Animation")
-		layout.prop_menu_enum(rd, "display_mode", text="Display Mode")
 
 		layout.separator()
 
 		layout.operator("render.opengl", text="Viewport Render Image", icon='RENDER_STILL')
-		props = layout.operator("render.opengl", text="Viewport Render Animation", icon='RENDER_ANIMATION')
+		layout.operator("render.opengl", text="Viewport Render Animation", icon='RENDER_ANIMATION').animation = True
+		if bpy.app.version > (2, 91):
+			props = layout.operator("render.opengl",text="Viewport Render Keyframes",icon='RENDER_ANIMATION',
+                                )
+			props.animation = True
+			props.render_keyed_only = True
 
 		layout.separator()
 		if bpy.app.version > (2, 81):
 			layout.operator("screen.info_log_show", text="Display Log Info", icon='INFO')
+			if not sys.platform == 'darwin':
+				layout.operator("wm.console_toggle", text="Toggle System Console", icon='CONSOLE')
+		
+		layout.separator()
+		layout.operator("wm.qs_path_open", text="Reveal Blend file", icon='FILE_FOLDER')
 
 
 def get_hotkey_entry_item(km, kmi_name, kmi_value, properties):
@@ -475,13 +630,20 @@ def get_hotkey_entry_item(km, kmi_name, kmi_value, properties):
 	returns hotkey of specific type, with specific properties.name (keymap is not a dict, so referencing by keys is not enough
 	if there are multiple hotkeys!)
 	'''
+	
 	for i, km_item in enumerate(km.keymap_items):
 		if km.keymap_items.keys()[i] == kmi_name:
+			# Use for Switch Engine, PopupMenu & PieMenu
 			if properties == 'name':
 				if km.keymap_items[i].properties.name == kmi_value:
 					return km_item
 			try:
 				if km.keymap_items[i].properties.layoutName == kmi_value:
+					# print(properties)
+					# print(km.keymap_items[i].properties.layoutName)
+					# print(km.keymap_items[i].properties.wslayoutMenu)
+					# print("\n\n")
+					# print("%s - %s" % (km.keymap_items[i].properties.layoutName, kmi_value))
 					return km_item
 			except:
 				pass
@@ -500,9 +662,8 @@ class QS_MT_WorkspaceSwitchPieMenu(Menu):
 		wm = bpy.context.window_manager
 		kc = wm.keyconfigs.user
 		km = kc.keymaps['Screen']
-
-		icons = [('Layout','VIEW3D'),('Modeling','VIEW3D'),('Sculpting','SCULPTMODE_HLT'),('UV Editing','GROUP_UVS'),('Texture Paint','IMAGE'),('Shading','SHADING_RENDERED'),('Animation','RENDER_ANIMATION'),('Rendering','RENDER_STILL'),('Compositing','NODE_COMPOSITING'),('Scripting','CONSOLE')]
-
+		
+		icons = [('Layout','VIEW3D'),('Modeling','VIEW3D'),('Sculpting','SCULPTMODE_HLT'),('UV Editing','UV'),('Texture Paint','IMAGE'),('Shading','SHADING_RENDERED'),('Animation','RENDER_ANIMATION'),('Rendering','RENDER_STILL'),('Compositing','NODE_COMPOSITING'),('Scripting','CONSOLE'),('Default','WORKSPACE')]
 
 		for i in range(0,8):
 			kmi = get_hotkey_entry_item(km, 'qs.workspace_set_layout', 'WorkspaceSwitcher'+str(i), 'layoutname')
@@ -511,8 +672,10 @@ class QS_MT_WorkspaceSwitchPieMenu(Menu):
 					if kmi.properties.wslayoutMenu in icons[k][0]:
 						icon = icons[k][1]
 						break
-					else:
+					elif kmi.properties.wslayoutMenu == "Preferences":
 						icon = 'PREFERENCES'
+					else:
+						icon = 'WORKSPACE'
 				if i == 3:
 					pie = layout.menu_pie()
 					split = pie.split()
@@ -521,7 +684,7 @@ class QS_MT_WorkspaceSwitchPieMenu(Menu):
 					#row = col.row(align=True)
 					col.scale_y=1.5
 					col.scale_x=1.5
-					#col.prop(scene,"qsStore3dView")
+					col.prop(scene,"qsStore3dView")
 					col.prop(scene,"qsKeepMode")
 					col.operator("qs.workspace_set_layout", text='{}'.format(kmi.properties.wslayoutMenu),icon=icon).wslayoutMenu=kmi.properties.wslayoutMenu
 				else:
@@ -537,7 +700,7 @@ class QS_MT_WorkspaceSwitchMenu(Menu):
 		layout = self.layout
 		scene = context.scene
 
-		#layout.prop(scene,"qsStore3dView")
+		layout.prop(scene,"qsStore3dView")
 		layout.prop(scene,"qsKeepMode")
 		layout.separator()
 
@@ -546,36 +709,38 @@ class QS_MT_WorkspaceSwitchMenu(Menu):
 		km = kc.keymaps['Screen']
 
 		## Alphabetical order
-		#for i in range(0,len(avail_workspaces(self, context))):
-		#	layout.operator("qs.workspace_set_layout", text='{}'.format(avail_workspaces(self, context)[i][1]), icon='SEQ_SPLITVIEW').wslayoutMenu=avail_workspaces(self, context)[i][1]
+		#for i in range(0,len(get_names_workspaces(self, context))):
+		#	layout.operator("qs.workspace_set_layout", text='{}'.format(get_names_workspaces(self, context)[i][1]), icon='SEQ_SPLITVIEW').wslayoutMenu=get_names_workspaces(self, context)[i][1]
 
-		icons = [('Layout','VIEW3D'),('Modeling','VIEW3D'),('Sculpting','SCULPTMODE_HLT'),('UV Editing','GROUP_UVS'),('Texture Paint','IMAGE'),('Shading','SHADING_RENDERED'),('Animation','RENDER_ANIMATION'),('Rendering','RENDER_STILL'),('Compositing','NODE_COMPOSITING'),('Scripting','CONSOLE')]
+		icons = [('Layout','VIEW3D'),('Modeling','VIEW3D'),('Sculpting','SCULPTMODE_HLT'),('UV Editing','GROUP_UVS'),('Texture Paint','IMAGE'),('Shading','SHADING_RENDERED'),('Animation','RENDER_ANIMATION'),('Rendering','RENDER_STILL'),('Compositing','NODE_COMPOSITING'),('Scripting','CONSOLE'),('Default','WORKSPACE')]
 
 		## Custom order
-		for i in range(0,len(avail_workspaces(self, context))):
+		for i in range(0,len(get_names_workspaces(self, context))):
 			kmi = get_hotkey_entry_item(km, 'qs.workspace_set_layout', 'WorkspaceSwitcher'+str(i), 'layoutname')
 			if not kmi == None:
 				for k in range(0,len(icons)):
 					if kmi.properties.wslayoutMenu in icons[k][0]:
 						icon = icons[k][1]
 						break
-					else:
+					elif kmi.properties.wslayoutMenu == "Preferences":
 						icon = 'PREFERENCES'
+					else:
+						icon = 'WORKSPACE'
 
 				layout.operator("qs.workspace_set_layout", text='{}'.format(kmi.properties.wslayoutMenu),icon=icon).wslayoutMenu=kmi.properties.wslayoutMenu
 		layout.separator()
 
-def getView(context):
-	for area in bpy.context.screen.areas:
-		if area.type == 'VIEW_3D':
-			qssw = context.scene.qsStoreView
-			rv3d = area.spaces[0].region_3d
-			rv3d.view_location = viewLoc#float(qssw.viewLoc[0]).strip(", ") #,float(qssw.viewLoc[1]),float(qssw.viewLoc[2]))
-			rv3d.view_distance = distance#float(qssw.distance)
-			rv3d.view_rotation = rotation#float(qssw.rotation)
-			rv3d.view_matrix = matrix#float(qssw.rotation)
+# def getView(context):
+# 	for area in bpy.context.screen.areas:
+# 		if area.type == 'VIEW_3D':
+# 			qssw = context.scene.qsStoreView
+# 			rv3d = area.spaces[0].region_3d
+# 			rv3d.view_location = qssw.viewLoc#float(qssw.viewLoc[0]).strip(", ") #,float(qssw.viewLoc[1]),float(qssw.viewLoc[2]))
+# 			rv3d.view_distance = qssw.distance#float(qssw.distance)
+# 			rv3d.view_rotation = qssw.rotation#float(qssw.rotation)
+# 			# rv3d.view_matrix = matrix#float(qssw.rotation)
 
-			return
+# 			return
 
 
 class QS_OT_SetWorkspace(Operator):
@@ -583,15 +748,25 @@ class QS_OT_SetWorkspace(Operator):
 	bl_idname="qs.workspace_set_layout"
 	bl_label="Switch to Workspace"
 
-	wslayoutMenu: bpy.props.EnumProperty(name = "Workspace", items = avail_workspaces)
+	wslayoutMenu: bpy.props.EnumProperty(name = "Workspace", items = get_names_workspaces) #, update=update_names_workspaces)
 	layoutName: bpy.props.StringProperty()
 
 	def execute(self,context):
 		scene = context.scene
-		ws = context.workspace
+		ws = context.workspace # returns name before we set it, this is
+		
+		# from MACHIN3tools
+		wsp = bpy.data.workspaces.get(self.wslayoutMenu)
+		view = self.getView(context, wsp)
 
 		if self.wslayoutMenu == "Preferences":
 			bpy.ops.screen.userpref_show('INVOKE_DEFAULT')
+			# Go to Quickswitch addon
+			bpy.context.preferences.active_section = 'ADDONS'
+			bpy.context.window_manager.addon_search = 'quickswitch'
+			# force panel redraw
+			context.area.tag_redraw()
+
 			return{'FINISHED'}
 		else:
 			if scene.qsKeepMode:
@@ -609,21 +784,203 @@ class QS_OT_SetWorkspace(Operator):
 					wsN = 'UV_Editing'
 				if wsN == 'Texture Paint':
 					wsN = 'Texture_Paint'
-
+			
+				# Set interaction mode
 				ws.object_mode = scene.qsWSsmode[wsN]
-			#try:
-			if scene.qsStore3dView:
-				#items = scene.qsStoreView
-				getView(context)
+
+			# if not ws.name in scene.qsWSsmode:
+			# 	print("default mode")
+			# # else:
+			# # 	# Revert to default object mode, if default names are not used or workspace are cahnged after opening
+			# 	ws.object_mode = scene.qsWSsmode['Default']
+			# # 	print(self.wslayoutMenu)
+			# # 	print(self.layoutName)
+
+			# wsn = context.workspace
+			# ViewPort(wsn)
 			bpy.context.window.workspace = bpy.data.workspaces[self.wslayoutMenu]
-			return{'FINISHED'}
-			#except:
-				# except layout doesn't exists
-				#self.report({'INFO'}, 'Workspace [{}] doesn\'t exist! Create it or pick another in addon settings.'.format(self.layoutName))
+			
+			if scene.qsStore3dView:
+				# items = scene.qsStoreView
+				# print("Viewloc: %s" % items.viewLoc)
+				self.setView(context, wsp, view)
+
+		return {'FINISHED'}
+
+	def getView(self,context, wsp):
+		""" Return position, rotation data about a given view for the first space attached to it """
+		global viewLoc, distance, matrix, rotation
+
+		view = {}
+		
+		for i, area in enumerate(context.screen.areas):
+			if area.type == 'VIEW_3D':
+				for space in area.spaces:
+					if space.type == 'VIEW_3D':
+						# wsp = context.window.workspace
+						rv3d = space.region_3d
+						
+						# from MACHIN3tools
+						view["viewLoc"] = rv3d.view_location
+						view["distance"] = rv3d.view_distance
+						view["rotation"] = rv3d.view_rotation
+						# view["matrix"] = rv3d.view_matrix
+						view['view_perspective'] = rv3d.view_perspective
+
+						view['is_perspective'] = rv3d.is_perspective
+						view['is_side_view'] = rv3d.is_orthographic_side_view
+			# Return none for non 3dviews, otherwise error
+			else:
+				view["viewLoc"] = None
+
+		# return view if rv3d.view_perspective != 'CAMERA' else None
+		return view
+
+	def setView(self,context, workspace, view):
+		for screen in workspace.screens:
+			for area in screen.areas:
+				if area.type == 'VIEW_3D':
+					for space in area.spaces:
+						if space.type == 'VIEW_3D':
+							rv3d = space.region_3d
+							qssw = context.scene.qsStoreView
+							# rv3d = area.spaces[0].region_3d
+							# rv3d.view_location = qssw.viewLoc#float(qssw.viewLoc[0]).strip(", ") #,float(qssw.viewLoc[1]),float(qssw.viewLoc[2]))
+							# rv3d.view_distance = qssw.distance#float(qssw.distance)
+							# rv3d.view_rotation = qssw.rotation#float(qssw.rotation)
+							# rv3d.view_matrix = matrix#float(qssw.rotation)
+							
+							# If we come from a none 3dview skip it
+							# print(view["viewLoc"])
+							if view["viewLoc"] != None:
+								# from MACHIN3tools
+								rv3d.view_location = view["viewLoc"] #float(qssw.viewLoc[0]).strip(", ") #,float(qssw.viewLoc[1]),float(qssw.viewLoc[2]))
+								rv3d.view_distance = view["distance"] #float(qssw.distance)
+								rv3d.view_rotation = view["rotation"] #float(qssw.rotation)
+
+								# don't set camera views
+								if rv3d.view_perspective != 'CAMERA':
+									rv3d.view_perspective = view['view_perspective']
+
+									rv3d.is_perspective = view['is_perspective']
+									rv3d.is_orthographic_side_view = view['is_side_view']
+							else:
+								print("## QS > Skipping #DV Matching Workspace")
+							return
+
+
+
+#######################################################
+## 3DVIEW VIEW SETTINGS
+class QS_PG_ViewData(PropertyGroup):
+    name = StringProperty()
+    lens : FloatProperty()
+    clip_start : FloatProperty()
+    clip_end : FloatProperty()
+
+def QS_store_3dview_data(self, context):
+	view3d = context.space_data
+	# context = bpy.context
+	scene = context.scene
+	qs3dvd = scene.qs3DViewData
+
+	qs3dvd.lens = view3d.lens
+	qs3dvd.clip_start = view3d.clip_start
+	qs3dvd.clip_end = view3d.clip_end
+	if scene.qsSync3dView:
+		for ws in bpy.data.workspaces:
+			for area in bpy.data.workspaces[ws.name].screens[0].areas:
+				if area.type == 'VIEW_3D':
+					viewd = area.spaces.active
+					viewd.lens = qs3dvd.lens
+					viewd.clip_start = qs3dvd.clip_start
+					viewd.clip_end = qs3dvd.clip_end
+		
+
+class QD_OT_Sync_ViewData(Operator):
+	"""Syncs all 3DViews view settings"""
+	bl_idname = "qs.sync_viewdata"
+	bl_label = "Sync Viewports"
+	bl_options = {"REGISTER", "UNDO"}
+
+	def execute(self, context):
+		view3d = context.space_data
+		for ws in bpy.data.workspaces:
+			for area in bpy.data.workspaces[ws.name].screens[0].areas:
+				if area.type == 'VIEW_3D':
+					viewd = area.spaces.active
+					viewd.lens = view3d.lens
+					viewd.clip_start = view3d.clip_start
+					viewd.clip_end = view3d.clip_end
+
+		return {"FINISHED"}	
+
+
+class QD_OT_Reset_ViewData(Operator):
+	"""Resets all 3DViews view settings"""
+	bl_idname = "qs.reset_viewdata"
+	bl_label = "Reset Viewports"
+	bl_options = {"REGISTER", "UNDO"}
+
+	def execute(self, context):
+		for ws in bpy.data.workspaces:
+			for area in bpy.data.workspaces[ws.name].screens[0].areas:
+				if area.type == 'VIEW_3D':
+					viewd = area.spaces.active
+					viewd.lens = 50
+					viewd.clip_start = 0.1
+					viewd.clip_end = 1000
+
+		return {"FINISHED"}	
+					
+
+bpy.types.Scene.qsSync3dView = bpy.props.BoolProperty(
+	name = "Sync Viewports", 
+	default=False, 
+	description="Syncs all 3DViews view settings",
+	update=QS_store_3dview_data)
+
+
+# 
+# Function: Opens folder current blend file 
+# 
+class QS_OT_path_open(bpy.types.Operator):
+	"Open a path in a file browser"
+	bl_idname = "wm.qs_path_open"
+	bl_label = ""
+
+	filepath : StringProperty(name="File Path", maxlen= 1024)
+
+	@classmethod
+	def poll(cls, context):
+		return os.path.dirname(bpy.data.filepath)
+
+	def execute(self, context):
+		filepath = os.path.dirname(bpy.data.filepath)
+		
+		if not os.path.exists(filepath):
+			self.report({'ERROR'}, "File '%s' not found" % filepath)
+			return {'CANCELLED'}
+		print(filepath)
+		if sys.platform == 'win32':
+			# subprocess.Popen(['start', filepath], shell= True)
+			# https://stackoverflow.com/questions/281888/open-explorer-on-a-file
+			subprocess.Popen(r'explorer /open,"'+filepath+'\"')
+		elif sys.platform == 'darwin':
+			subprocess.Popen(['open', "-R", filepath])
 
 		return {'FINISHED'}
 
 
+def ui_add_menu(self, context):
+	# scene = context.scene
+	col = self.layout.column(align=True)
+	row = col.row(align=True)
+	row.label(text='Sync views')
+	row.operator(QD_OT_Sync_ViewData.bl_idname, icon="UV_SYNC_SELECT",text="")
+	# row.operator(QD_OT_Sync_ViewData.bl_idname, icon="UV_SYNC_SELECT")
+	row.operator(QD_OT_Reset_ViewData.bl_idname, icon="LOOP_BACK", text="")
+    
 ########################################################
 
 
@@ -632,7 +989,7 @@ class QS_PT_AddonPreferences(AddonPreferences):
 	""" Preference Settings Addin Panel"""
 	bl_idname = __name__
 
-	qsMenus: bpy.props.EnumProperty(name = "QuickSwitch Options", items = [("Workspaces","Workspaces","Workspaces"),("Render Menu","Render Menu","Render Menu")])
+	qsMenus: bpy.props.EnumProperty(name = "QuickSwitch Options", items = [("Workspaces","Workspaces","Workspaces"),("Render_Menu","Render Menu","Render Menu")])
 
 	def draw(self, context):
 		layout = self.layout
@@ -677,7 +1034,7 @@ class QS_PT_AddonPreferences(AddonPreferences):
 			split = box.split()
 			col = split.column()
 			col.label(text='Set Workspaces:')
-			for i in range(0,len(avail_workspaces(self, context))):
+			for i in range(0,len(get_names_workspaces(self, context))):
 	#			if km.keymap_items.keys()[i] == 'Switch to Workspace':
 				kmi = get_hotkey_entry_item(km, 'qs.workspace_set_layout', 'WorkspaceSwitcher'+str(i), 'layoutname')
 				if not kmi == None:
@@ -688,7 +1045,7 @@ class QS_PT_AddonPreferences(AddonPreferences):
 						col.label(text="restore hotkeys from interface tab")
 			col.label(text='Set each shortcut in the dropdown menu named "Workspace"')
 
-		if getattr(self,"qsMenus") in ("Render Menu"):
+		if getattr(self,"qsMenus") in ("Render_Menu"):
 			wm = bpy.context.window_manager
 			kc = wm.keyconfigs.user
 			km = kc.keymaps['Screen']
@@ -708,48 +1065,62 @@ class QS_PT_AddonPreferences(AddonPreferences):
 
 
 #Classes for register and unregister
-classes = (
+classes = [
+	QS_workspaceStore,
 	QS_defaultWSSmodes,
-	#QS_Store3DView,
+	QS_Store3DView,
+	# QS_StoreView,
 	#AddPresetQuickSwitch,
 	#QS_PT_presets,
 	#QS_OT_AddonUpdater,
+	QS_PG_ViewData,
+	QD_OT_Sync_ViewData,
+	QD_OT_Reset_ViewData,
 	QS_PT_AddonPreferences,
 	QS_MT_QuickSwitchEngine,
 	QS_MT_WorkspaceSwitchPieMenu,
 	QS_MT_WorkspaceSwitchMenu,
 	QS_OT_SetWorkspace,
-	)
+	QS_OT_path_open,
+]
 
 def register():
 	for cls in classes:
 		bpy.utils.register_class(cls)
 
+	bpy.types.Scene.qsWSstore = PointerProperty(type=QS_workspaceStore)
 	bpy.types.Scene.qsWSsmode = PointerProperty(type=QS_defaultWSSmodes)
-	#bpy.types.Scene.qsStoreView = PointerProperty(type=QS_Store3DView)
+	bpy.types.Scene.qsStoreView = PointerProperty(type=QS_Store3DView)
+	bpy.types.Scene.qs3DViewData = PointerProperty(type=QS_PG_ViewData)
 
 	# hotkey setup
 	add_hotkey()
 
+	# Check workspace names and interaction mode
 	bpy.app.handlers.depsgraph_update_pre.append(on_scene_update)
-	#bpy.app.handlers.depsgraph_update_pre.append(on_ws_switch)
+	# bpy.app.handlers.depsgraph_update_pre.append(on_ws_switch)
+	bpy.types.VIEW3D_PT_view3d_properties.prepend(ui_add_menu)
 
 
 
 def unregister():
 	bpy.app.handlers.depsgraph_update_pre.remove(on_scene_update)
-	#bpy.app.handlers.depsgraph_update_pre.remove(on_ws_switch)
-
+	# bpy.app.handlers.depsgraph_update_pre.remove(on_ws_switch)
+	bpy.types.VIEW3D_PT_view3d_properties.remove(ui_add_menu)
+	
 	# handle the keymap
 	for km, kmi in addon_keymaps:
+		print("QS: %s - %s" % (km,kmi))
 		km.keymap_items.remove(kmi)
 	addon_keymaps.clear()
-
 	for cls in reversed(classes):
 		bpy.utils.unregister_class(cls)
 
 	del bpy.types.Scene.qsWSsmode
-	#del bpy.types.Scene.qsStoreView
+	del bpy.types.Scene.qsStoreView
+	# del bpy.types.Scene.qs3DViewData
+
+
 
 if __name__ == "__main__":
 	register()